Cabin task is added
Added a task which done by holding down a button. After determined time an object is spawned in player's hand and older objects meshes are disabled. Then the player must put this object to the predetermined pivot locations. I also made some adjustments and decorations to the scene and added a music.  Feels a lot nicer now.